dm: platform_data: atmel_serial: fix build warning on 64-bit platforms
Pointer size cannot be assumed to be 32-bit, so use uintptr_t instead of
uint32_t.

Fixes the below build warning on 64-bit builds:
drivers/serial/atmel_usart.c: In function ‘atmel_serial_probe’:
drivers/serial/atmel_usart.c:275:23: warning: cast to pointer from integer of different size [-Wint-to-pointer-cast]
  275 |         priv->usart = (atmel_usart3_t *)plat->base_addr;

+++ include/dm/platform_data/atmel_serial.h
@@ -8,7 +8,7 @@
 
 /* Information about a serial port */
 struct atmel_serial_plat {
-	uint32_t base_addr;
+	uintptr_t base_addr;
 };
 
 #endif
